requirements.txt
python-dotenv>=1.0.0
groq>=0.4.0

